fix(fam65xx): correct BRK interrupt priority and 65816 stack pointer handling
- Change BRK interrupt condition from checking if no interrupt is active to allowing BRK to override COP/IRQ interrupts
- Only NMI and RESET interrupts (higher priority) can now override BRK
- Correct stack pointer type from uint8_t to uint16_t in JSON parser for 65816 native mode compatibility
- Update comments to clarify interrupt priority behavior

+++ code/cpp/tests/json_parser.cpp
@@ -330,7 +330,7 @@ bool json_parse_cpu_state(const char* json, const char* state_name, cpu_state_t*
     
     // Parse registers - preserve full 16-bit values for 65816 compatibility
     state->pc = (uint16_t)json_parse_number(state_obj, "pc");
-    state->s = (uint8_t)json_parse_number(state_obj, "s");   // Stack pointer is always 8-bit
+    state->s = (uint16_t)json_parse_number(state_obj, "s");  // Stack pointer: 8-bit for 6502, 16-bit for 65816 native mode
     state->a = (uint16_t)json_parse_number(state_obj, "a");  // Accumulator: 8-bit for 6502, 16-bit for 65816
     state->x = (uint16_t)json_parse_number(state_obj, "x");  // X register: 8-bit for 6502, 16-bit for 65816
     state->y = (uint16_t)json_parse_number(state_obj, "y");  // Y register: 8-bit for 6502, 16-bit for 65816
